% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TherapyClass.R
\name{dpd_class}
\alias{dpd_class}
\title{Therapeutic Class}
\usage{
dpd_class(id)
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble}.
}
\description{
A drug's Therapeutic Classification (Class) is assigned according to its main
therapeutic use.
}
