% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Company.R
\name{dpd_company}
\alias{dpd_company}
\title{Company}
\usage{
dpd_company(ids)
}
\arguments{
\item{ids}{Vector of company codes.}
}
\value{
A \code{tibble} with columns:
\itemize{
\item \code{city_name}: City name.
\item \code{company_code}: Code assigned to a company.
\item \code{company_name}: Company name.
\item \code{company_type}: Company type.
\item \code{country_name}: Country name.
\item \code{post_office_box}: P.O. box number (if applicable).
\item \code{postal_code}: Postal code.
\item \code{province_name}: Province name.
\item \code{street_name}: Street name.
\item \code{suite_number}: Suite number.
}
}
\description{
Company information associated with a company code.
}
\examples{
dpd_company(10825)

}
