% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosageForm.R
\name{dpd_dosage}
\alias{dpd_dosage}
\alias{dpd_dosage_all}
\title{Dosage Form}
\usage{
dpd_dosage(id)

dpd_dosage_all()
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble}.
}
\description{
The dosage form is the form of presentation in which the product is supplied,
for example, tablet, capsule, powder, etc.
}
\details{
A product can have more than one dosage form when it is a kit
(e.g. tablet, capsule).
}
