% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Packaging.R
\name{dpd_packaging}
\alias{dpd_packaging}
\title{Packaging}
\usage{
dpd_packaging(id)
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble}.
}
\description{
Information about available package sizes for each drug product.
}
