% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmStandard.R
\name{dpd_pharm_standard}
\alias{dpd_pharm_standard}
\title{Pharmaceutical Standard}
\usage{
dpd_pharm_standard(id, lang = c("en", "fr"))
}
\arguments{
\item{id}{Drug product code.}

\item{lang}{Language of the response. One of \code{"en"} or \code{"fr"} for English or
French, respectively.}
}
\value{
A \code{tibble} with columns:
\itemize{
\item \code{drug_code}: code assigned to each drug product.
\item \code{pharmaceutical_std}: Pharmaceutical_std
}
}
\description{
The standard to which a drug product is manufactured and represented.
}
\examples{
dpd_pharm_standard(11534)
}
