% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmStandard.R
\name{dpd_pharm_std}
\alias{dpd_pharm_std}
\title{Pharmaceutical Standard}
\usage{
dpd_pharm_std(id)
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble}.
}
\description{
The standard to which a drug product is manufactured and represented.
}
