% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schedule.R
\name{dpd_schedule}
\alias{dpd_schedule}
\alias{dpd_schedule_all}
\title{Schedule}
\usage{
dpd_schedule(id)

dpd_schedule_all()
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble}.
}
\description{
Each drug is assigned one or more schedules, according to the Food and Drug
Regulations, and the Controlled Drugs and Substances Act.
}
