% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VetSpecies.R
\name{dpd_species}
\alias{dpd_species}
\title{Veterinary Species}
\usage{
dpd_species(id)
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble}.
}
\description{
The type of species for a certain veterinary product (e.g. sheep, dog,
cattle, and poultry).
}
