% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductStatus.R
\name{dpd_status}
\alias{dpd_status}
\title{Product Status}
\usage{
dpd_status(id)
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble}.
}
\description{
The statuses listed below are a direct representation of the status available
in DPD Online Query and the description of each status is summarized below.
}
