% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VetSpecies.R
\name{dpd_vet_species}
\alias{dpd_vet_species}
\title{Veterinary Species}
\usage{
dpd_vet_species(id, lang = c("en", "fr"))
}
\arguments{
\item{id}{Drug product code.}

\item{lang}{Language of the response. One of \code{"en"} or \code{"fr"} for English or
French, respectively.}
}
\value{
A \code{tibble} with columns:
\itemize{
\item \code{drug_code}: Code assigned to each drug product.
\item \code{vet_species_name}: Animal species.
}
}
\description{
The type of species for a certain veterinary product (e.g. sheep, dog,
cattle, and poultry).
}
\examples{
dpd_vet_species(13755)
}
