% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpdr-package.R
\docType{package}
\name{dpdr-package}
\alias{dpdr}
\alias{dpdr-package}
\title{dpdr: Interface to Health Canada Drug Product Database API}
\description{
This package provides a programmatic interface to the Health Canada Drug Product Database REST API for querying specific information on drugs approved for use in Canada.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattwarkentin/dpdr}
  \item \url{https://mattwarkentin.github.io/dpdr/}
  \item Report bugs at \url{https://github.com/mattwarkentin/dpdr/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
