% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActiveIngredient.R
\name{dpd_ai_id}
\alias{dpd_ai_id}
\alias{dpd_ai_name}
\alias{dpd_ai_all}
\title{Active Ingredient}
\usage{
dpd_ai_id(id)

dpd_ai_name(name)

dpd_ai_all()
}
\arguments{
\item{id}{Drug product code.}

\item{name}{Active ingredient name.}
}
\value{
A \code{tibble}.
}
\description{
An Active Ingredient is any component that has medicinal properties, and
supplies pharmacological activity or other direct effect in the diagnosis,
cure, mitigation, treatment or prevention of disease, or to affect the
structure or any function of the body of a human or an animal.
}
