% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TherapyClass.R
\name{dpd_class}
\alias{dpd_class}
\title{Therapeutic Class}
\usage{
dpd_class(ids)
}
\arguments{
\item{ids}{Vector of drug product codes.}
}
\value{
A \code{tibble} with columns:
\itemize{
\item \code{drug_code}: Code assigned to each drug product.
\item \code{tc_ahfs}: Description of the American Hospital Formulary Service.
\item \code{tc_ahfs_number}: American Hospital Formulary Service code.
\item \code{tc_atc}: Description of the Anatomical Therapeutic Chemical Code.
\item \code{tc_atc_number}: Anatomical Therapeutic Chemical Code.
}
}
\description{
A drug's Therapeutic Classification (Class) is assigned according to its main
therapeutic use.
}
\examples{
dpd_class(10564)

}
