% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Company.R
\name{dpd_company}
\alias{dpd_company}
\title{Company}
\usage{
dpd_company(id)
}
\arguments{
\item{id}{Company code.}
}
\value{
A \code{tibble}.
}
\description{
In the DPD, a product may have several companies associated with them. Each
product will have a DIN Owner associated with it.
}
