% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosageForm.R
\name{dpd_dosage}
\alias{dpd_dosage}
\alias{dpd_dosage_all}
\title{Dosage Form}
\usage{
dpd_dosage(ids)

dpd_dosage_all()
}
\arguments{
\item{ids}{Vector of drug product codes.}
}
\value{
A \code{tibble} with columns:
\itemize{
\item \code{drug_code}: Code assigned to each drug product.
\item \code{pharmaceutical_form_code}: Code assigned to a dosage form.
\item \code{pharmaceutical_form_name}: Dosage form.
}
}
\description{
The dosage form is the form of presentation in which the product is supplied,
for example, tablet, capsule, powder, etc.
}
\details{
A product can have more than one dosage form when it is a kit
(e.g. tablet, capsule).
}
\examples{
dpd_dosage(10846)

}
