% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugProduct.R
\name{dpd_drug_id}
\alias{dpd_drug_id}
\alias{dpd_drug_din}
\alias{dpd_drug_brand}
\alias{dpd_drug_status}
\alias{dpd_drug_all}
\alias{dpd_drug_dins}
\title{Drug Product}
\usage{
dpd_drug_id(id)

dpd_drug_din(din)

dpd_drug_brand(brand)

dpd_drug_status(status)

dpd_drug_all()

dpd_drug_dins(name)
}
\arguments{
\item{id}{Drug product code.}

\item{din}{Drug Identification Number (DIN).}

\item{brand}{Brand name.}

\item{status}{Drug product status.}

\item{name}{Active ingredient name.}
}
\value{
A \code{tibble}.
}
\description{
Basic information about the product, such as brand name and Drug
Identification Number.
}
