% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmStandard.R
\name{dpd_pharm_std}
\alias{dpd_pharm_std}
\title{Pharmaceutical Standard}
\usage{
dpd_pharm_std(ids)
}
\arguments{
\item{ids}{Vector of drug product codes.}
}
\value{
A \code{tibble} with columns:
\itemize{
\item \code{drug_code}: code assigned to each drug product.
\item \code{pharmaceutical_std}: Pharmaceutical_std
}
}
\description{
The standard to which a drug product is manufactured and represented.
}
\examples{
dpd_pharm_std(11534)

}
