% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VetSpecies.R
\name{dpd_species}
\alias{dpd_species}
\title{Veterinary Species}
\usage{
dpd_species(id)
}
\arguments{
\item{id}{Drug product code.}
}
\value{
A \code{tibble} with columns:
\itemize{
\item \code{drug_code}: Code assigned to each drug product.
\item \code{vet_species_name}: Animal species.
}
}
\description{
The type of species for a certain veterinary product (e.g. sheep, dog,
cattle, and poultry).
}
\examples{
dpd_species(13755)

}
