% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpdr-package.R
\docType{package}
\name{dpdr-package}
\alias{dpdr}
\alias{dpdr-package}
\title{dpdr: Interface to Health Canada Drug Product Database API}
\description{
Provides a programmatic interface to the Health Canada Drug Product Database Application Programming Interface (API) for querying specific information on drugs approved for use in Canada.
}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{warkentin@lunenfeld.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
