% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpdr-package.R
\docType{package}
\name{dpdr-package}
\alias{dpdr}
\alias{dpdr-package}
\title{dpdr: Interface to Health Canada Drug Product Database API}
\description{
This package provides a programmatic interface to Health Canada's Drug Product Database (DPD) REST API for querying information about drugs approved for use in Canada. More information on the DPD can be found in the API guide (\url{https://health-products.canada.ca/api/documentation/dpd-documentation-en.html}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattwarkentin/dpdr}
  \item \url{https://mattwarkentin.github.io/dpdr/}
  \item Report bugs at \url{https://github.com/mattwarkentin/dpdr/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID}) [copyright holder]

}
\keyword{internal}
